Struct leo_imports::parser::import_parser::ImportParser

pub struct ImportParser {
    program_path: PathBuf,
    partial_imports: IndexSet<String>,
    imports: IndexMap<String, Program>,
    pub imports_map: IndexMap<String, String>,
}

Stores imported packages.

A program can import one or more packages. A package can be found locally in the source directory, foreign in the imports directory, or part of the core package list.

Fields
program_path: PathBufpartial_imports: IndexSet<String>imports: IndexMap<String, Program>imports_map: IndexMap<String, String>
Implementations

impl ImportParser

pub(crate) fn parse_import_file(
package: &DirEntry,
span: &Span
) -> Result<Program>

Returns a Leo syntax tree from a given package.

Builds an abstract syntax tree from the given file and then builds the Leo syntax tree.

impl ImportParser

pub fn new(program_path: PathBuf, imports_map: IndexMap<String, String>) -> Self

impl ImportParser

fn parse_package_access(
&mut self,
package: &DirEntry,
remaining_segments: &[&str],
span: &Span
) -> Result<Program>

pub(crate) fn parse_package(
&mut self,
path: PathBuf,
segments: &[&str],
span: &Span
) -> Result<Program>

Create the Leo syntax tree for an imported package.

Inserts the Leo syntax tree into the ImportParser.